+++ modules/admin/tests/data/fixtures/UserFixture.php
@@ -13,6 +13,7 @@ class UserFixture extends ActiveFixture
         return [
             'user1' => [
                 'id' => 1,
+                'title' => 1,
                 'firstname' => 'John',
                 'lastname' => 'Doe',
                 'email' => 'john@luya.io',
@@ -21,6 +22,7 @@ class UserFixture extends ActiveFixture
             ],
             'user2' => [
                 'id' => 2,
+                'title' => 2,
                 'firstname' => 'Jane',
                 'lastname' => 'Doe',
                 'email' => 'jane@luya.io',